pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marble-bust-training"
version = "0.1.0"
description = "Separated marble bust training pipelines"
requires-python = ">=3.12,<3.13"
dependencies = [
  "boto3>=1.35",
  "datasets>=3.0",
  "jinja2>=3.1",
  "pillow>=10.0",
  "pydantic>=2.9",
  "pyyaml>=6.0.2",
  "typer>=0.12",
]

[project.optional-dependencies]
test = ["pytest>=8.3"]
inference = [
  "accelerate>=1.0",
  "diffusers @ git+https://github.com/huggingface/diffusers.git",
  "peft>=0.13",
  "safetensors>=0.4",
  "sentencepiece>=0.2",
  "transformers>=4.57",
]

[tool.setuptools.packages.find]
include = ["core*", "backends*", "data*", "inference*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
